[checker] Fully merge global context building and optimization
