+++ samlang-core/checker/__tests__/global-typing-context-builder.test.ts
@@ -6,13 +6,20 @@ import {
 } from '../../ast/common-nodes';
 import type { SamlangModule, SourceClassDefinition } from '../../ast/samlang-nodes';
 import {
+  prettyPrintType,
+  SamlangIdentifierType,
   SourceExpressionFalse,
   SourceFunctionType,
   SourceId,
   SourceIdentifierType,
   SourceIntType,
 } from '../../ast/samlang-nodes';
-import { buildGlobalTypingContext } from '../global-typing-context-builder';
+import { createGlobalErrorCollector } from '../../errors';
+import {
+  buildGlobalTypingContext,
+  getFullyInlinedInterfaceContext,
+} from '../global-typing-context-builder';
+import { GlobalTypingContext, memberTypeInformationToString } from '../typing-context';
 
 const module0Reference = ModuleReference(['Module0']);
 const module1Reference = ModuleReference(['Module1']);
@@ -93,10 +100,14 @@ const testSources = ModuleReferenceCollections.mapOf(
 
 describe('global-typing-context-builder', () => {
   it('can handle imports and definitions', () => {
-    const actualGlobalTypingContext = buildGlobalTypingContext(testSources, {
-      classes: new Map(),
-      interfaces: new Map(),
-    });
+    const actualGlobalTypingContext = buildGlobalTypingContext(
+      testSources,
+      createGlobalErrorCollector().getErrorReporter(),
+      {
+        classes: new Map(),
+        interfaces: new Map(),
+      },
+    );
     expect(actualGlobalTypingContext.size).toBe(3);
 
     expect(actualGlobalTypingContext.get(module0Reference)).toStrictEqual({
@@ -198,5 +209,238 @@ describe('global-typing-context-builder', () => {
         ],
       ]),
     });
+  });
+
+  it('getFullyInlinedInterfaceContext tests', () => {
+    const globalTypingContext: GlobalTypingContext = ModuleReferenceCollections.hashMapOf([
+      ModuleReference.DUMMY,
+      {
+        interfaces: new Map([
+          [
+            'IUseNonExistent',
+            {
+              typeParameters: [
+                { name: 'A', bound: null },
+                { name: 'B', bound: null },
+              ],
+              extendsOrImplements: SourceIdentifierType(
+                DummySourceReason,
+                ModuleReference.DUMMY,
+                'not_exist',
+              ),
+              superTypes: [],
+              functions: new Map(),
+              methods: new Map(),
+            },
+          ],
+          [
+            'IBase',
+            {
+              typeParameters: [
+                { name: 'A', bound: null },
+                { name: 'B', bound: null },
+              ],
+              extendsOrImplements: null,
+              superTypes: [],
+              functions: new Map(),
+              methods: new Map([
+                [
+                  'm1',
+                  {
+                    isPublic: true,
+                    typeParameters: [{ name: 'C', bound: null }],
+                    type: SourceFunctionType(
+                      DummySourceReason,
+                      [
+                        SourceIdentifierType(DummySourceReason, ModuleReference.DUMMY, 'A'),
+                        SourceIdentifierType(DummySourceReason, ModuleReference.DUMMY, 'B'),
+                      ],
+                      SourceIdentifierType(DummySourceReason, ModuleReference.DUMMY, 'C'),
+                    ),
+                  },
+                ],
+              ]),
+            },
+          ],
+          [
+            'ILevel1',
+            {
+              typeParameters: [
+                { name: 'A', bound: null },
+                { name: 'B', bound: null },
+              ],
+              extendsOrImplements: SourceIdentifierType(
+                DummySourceReason,
+                ModuleReference.DUMMY,
+                'IBase',
+                [
+                  SourceIntType(DummySourceReason),
+                  SourceIdentifierType(DummySourceReason, ModuleReference.DUMMY, 'B'),
+                ],
+              ),
+              superTypes: [],
+              functions: new Map([
+                [
+                  'f1',
+                  {
+                    isPublic: true,
+                    typeParameters: [{ name: 'C', bound: null }],
+                    type: SourceFunctionType(
+                      DummySourceReason,
+                      [
+                        SourceIdentifierType(DummySourceReason, ModuleReference.DUMMY, 'A'),
+                        SourceIdentifierType(DummySourceReason, ModuleReference.DUMMY, 'B'),
+                      ],
+                      SourceIdentifierType(DummySourceReason, ModuleReference.DUMMY, 'C'),
+                    ),
+                  },
+                ],
+              ]),
+              methods: new Map(),
+            },
+          ],
+          [
+            'ILevel2',
+            {
+              typeParameters: [
+                { name: 'A', bound: null },
+                { name: 'B', bound: null },
+              ],
+              extendsOrImplements: SourceIdentifierType(
+                DummySourceReason,
+                ModuleReference.DUMMY,
+                'ILevel1',
+                [
+                  SourceIdentifierType(DummySourceReason, ModuleReference.DUMMY, 'A'),
+                  SourceIntType(DummySourceReason),
+                ],
+              ),
+              superTypes: [],
+              functions: new Map(),
+              methods: new Map([
+                [
+                  'm2',
+                  {
+                    isPublic: true,
+                    typeParameters: [{ name: 'C', bound: null }],
+                    type: SourceFunctionType(
+                      DummySourceReason,
+                      [
+                        SourceIdentifierType(DummySourceReason, ModuleReference.DUMMY, 'A'),
+                        SourceIdentifierType(DummySourceReason, ModuleReference.DUMMY, 'B'),
+                      ],
+                      SourceIdentifierType(DummySourceReason, ModuleReference.DUMMY, 'C'),
+                    ),
+                  },
+                ],
+              ]),
+            },
+          ],
+          [
+            'ICyclic1',
+            {
+              typeParameters: [],
+              extendsOrImplements: SourceIdentifierType(
+                DummySourceReason,
+                ModuleReference.DUMMY,
+                'ICyclic2',
+              ),
+              superTypes: [],
+              functions: new Map(),
+              methods: new Map(),
+            },
+          ],
+          [
+            'ICyclic2',
+            {
+              typeParameters: [],
+              extendsOrImplements: SourceIdentifierType(
+                DummySourceReason,
+                ModuleReference.DUMMY,
+                'ICyclic1',
+              ),
+              superTypes: [],
+              functions: new Map(),
+              methods: new Map(),
+            },
+          ],
+        ]),
+        classes: new Map(),
+      },
+    ]);
+
+    function inlinedContextFromType(idType: SamlangIdentifierType) {
+      const { functions, methods, superTypes } = getFullyInlinedInterfaceContext(
+        idType,
+        globalTypingContext,
+        createGlobalErrorCollector().getErrorReporter(),
+      );
+      const functionStrings: string[] = [];
+      const methodStrings: string[] = [];
+      for (const [name, info] of functions) {
+        functionStrings.push(memberTypeInformationToString(name, info));
+      }
+      for (const [name, info] of methods) {
+        methodStrings.push(memberTypeInformationToString(name, info));
+      }
+      return {
+        functions: functionStrings,
+        methods: methodStrings,
+        superTypes: superTypes.map((it) => prettyPrintType(it)),
+      };
+    }
+
+    expect(
+      inlinedContextFromType(
+        SourceIdentifierType(DummySourceReason, ModuleReference.DUMMY, 'I_not_exist'),
+      ),
+    ).toEqual({
+      functions: [],
+      methods: [],
+      superTypes: [],
+    });
+
+    expect(
+      inlinedContextFromType(
+        SourceIdentifierType(DummySourceReason, ModuleReference.DUMMY, 'IUseNonExistent'),
+      ),
+    ).toEqual({
+      functions: [],
+      methods: [],
+      superTypes: ['not_exist'],
+    });
+    expect(
+      inlinedContextFromType(SourceIdentifierType(DummySourceReason, ModuleReference.DUMMY, 'I')),
+    ).toEqual({
+      functions: [],
+      methods: [],
+      superTypes: [],
+    });
+    expect(
+      inlinedContextFromType(
+        SourceIdentifierType(DummySourceReason, ModuleReference.DUMMY, 'ILevel2'),
+      ),
+    ).toEqual({
+      functions: ['public f1<C>(A, B) -> C'],
+      methods: ['public m1<C>(int, int) -> C', 'public m2<C>(A, B) -> C'],
+      superTypes: ['IBase<int, int>', 'ILevel1<A, int>'],
+    });
+
+    const errorCollector = createGlobalErrorCollector();
+    const errorReporter = errorCollector.getErrorReporter();
+    getFullyInlinedInterfaceContext(
+      SourceIdentifierType(DummySourceReason, ModuleReference.DUMMY, 'ICyclic1'),
+      globalTypingContext,
+      errorReporter,
+    );
+    getFullyInlinedInterfaceContext(
+      SourceIdentifierType(DummySourceReason, ModuleReference.DUMMY, 'ICyclic2'),
+      globalTypingContext,
+      errorReporter,
+    );
+    expect(errorCollector.getErrors().map((it) => it.toString())).toEqual([
+      '__DUMMY__.sam:0:0-0:0: [CyclicTypeDefinition]: Type `ICyclic1` has a cyclic definition.',
+      '__DUMMY__.sam:0:0-0:0: [CyclicTypeDefinition]: Type `ICyclic2` has a cyclic definition.',
+    ]);
   });
 });
