+++ samlang-core/checker/global-typing-context-builder.ts
@@ -1,132 +1,375 @@
 /** Responsible for building the global typing environment as part of pre-processing phase. */
 
 import {
+  DummySourceReason,
   ModuleReference,
   ModuleReferenceCollections,
   SourceReason,
   Sources,
 } from '../ast/common-nodes';
 import {
+  isTheSameType,
+  SamlangFunctionType,
+  SamlangIdentifierType,
   SamlangModule,
   SourceClassDefinition,
+  SourceClassMemberDeclaration,
   SourceFunctionType,
   SourceIdentifierType,
   SourceInterfaceDeclaration,
 } from '../ast/samlang-nodes';
-import { checkNotNull } from '../utils';
+import type { GlobalErrorReporter } from '../errors';
+import { checkNotNull, HashMap, zip } from '../utils';
+import performTypeSubstitution from './type-substitution';
 import {
   ClassTypingContext,
   DEFAULT_BUILTIN_TYPING_CONTEXT,
   GlobalTypingContext,
+  InterfaceTypingContext,
   MemberTypeInformation,
   ModuleTypingContext,
 } from './typing-context';
 
-function buildInterfaceTypingContext({
-  typeParameters,
-  members,
-  extendsOrImplementsNode,
-}: SourceInterfaceDeclaration) {
-  const functions = new Map<string, MemberTypeInformation>();
-  const methods = new Map<string, MemberTypeInformation>();
-  members.forEach(({ name, isPublic, isMethod, type, typeParameters: memberTypeParameters }) => {
-    const typeInformation = {
-      isPublic,
-      typeParameters: memberTypeParameters.map((it) => ({ name: it.name.name, bound: it.bound })),
-      type,
-    };
-    if (isMethod) {
-      methods.set(name.name, typeInformation);
-    } else {
-      functions.set(name.name, typeInformation);
-    }
-  });
-  return {
-    typeParameters: typeParameters.map((it) => ({ name: it.name.name, bound: it.bound })),
-    extendsOrImplements: extendsOrImplementsNode ?? null,
-    superTypes: [],
-    functions,
-    methods,
-  };
-}
+type InterfaceInliningCollector = Array<{
+  readonly functions: Map<string, MemberTypeInformation>;
+  readonly methods: Map<string, MemberTypeInformation>;
+  readonly baseInterfaceType: SamlangIdentifierType | null;
+}>;
 
-function buildClassTypingContext(
-  moduleReference: ModuleReference,
-  classDefinition: SourceClassDefinition,
-): ClassTypingContext {
-  const { typeParameters, functions, methods } = buildInterfaceTypingContext(classDefinition);
-  const classType = SourceIdentifierType(
-    SourceReason(classDefinition.name.location, classDefinition.name.location),
-    moduleReference,
-    classDefinition.name.name,
-    classDefinition.typeParameters.map((it) =>
-      SourceIdentifierType(
-        SourceReason(it.location, it.location),
-        moduleReference,
-        it.name.name,
-        [],
-      ),
+function recursiveComputeInterfaceMembersChain(
+  interfaceType: SamlangIdentifierType,
+  globalTypingContext: GlobalTypingContext,
+  collector: InterfaceInliningCollector,
+  visited: HashMap<ModuleReference, Set<string>>,
+  errorReporter: GlobalErrorReporter,
+) {
+  const visitedTypesInModule = visited.get(interfaceType.moduleReference) ?? new Set();
+  if (visitedTypesInModule.has(interfaceType.identifier)) {
+    errorReporter.reportCyclicTypeDefinitionError(interfaceType);
+    return;
+  }
+  visited.set(interfaceType.moduleReference, visitedTypesInModule.add(interfaceType.identifier));
+  const interfaceContext = globalTypingContext
+    .get(interfaceType.moduleReference)
+    ?.interfaces?.get(interfaceType.identifier);
+  if (interfaceContext == null) return null;
+  const substitutionMapping = new Map(
+    zip(
+      interfaceContext.typeParameters.map((it) => it.name),
+      interfaceType.typeArguments,
     ),
   );
-  const { typeDefinition } = classDefinition;
-  const typeDefinitionReason = SourceReason(typeDefinition.location, typeDefinition.location);
-  if (typeDefinition.type === 'object') {
-    functions.set('init', {
-      isPublic: true,
-      typeParameters,
-      type: SourceFunctionType(
-        typeDefinitionReason,
-        typeDefinition.names.map((it) => checkNotNull(typeDefinition.mappings[it.name]).type),
-        classType,
-      ),
-    });
-  } else {
-    Object.entries(typeDefinition.mappings).forEach(([tag, { type }]) => {
-      functions.set(tag, {
-        isPublic: true,
+  const inlinedMethods = new Map(
+    Array.from(interfaceContext.methods, ([name, { isPublic, typeParameters, type }]) => [
+      name,
+      {
+        isPublic,
         typeParameters,
-        type: SourceFunctionType(typeDefinitionReason, [type], classType),
-      });
-    });
+        type: performTypeSubstitution(type, substitutionMapping) as SamlangFunctionType,
+      },
+    ]),
+  );
+  const baseInterfaceType =
+    interfaceContext.extendsOrImplements != null
+      ? (performTypeSubstitution(
+          interfaceContext.extendsOrImplements,
+          substitutionMapping,
+        ) as SamlangIdentifierType)
+      : null;
+  if (baseInterfaceType != null) {
+    recursiveComputeInterfaceMembersChain(
+      baseInterfaceType,
+      globalTypingContext,
+      collector,
+      visited,
+      errorReporter,
+    );
   }
-  return {
-    typeParameters,
-    typeDefinition,
-    extendsOrImplements: classDefinition.extendsOrImplementsNode ?? null,
-    superTypes: [],
-    functions,
-    methods,
-  };
+  collector.push({
+    functions: new Map(interfaceContext.functions),
+    methods: inlinedMethods,
+    baseInterfaceType,
+  });
 }
 
-function buildModuleTypingContext(
-  moduleReference: ModuleReference,
-  samlangModule: SamlangModule,
-): ModuleTypingContext {
-  return {
-    interfaces: new Map(
-      samlangModule.interfaces.map((declaration) => [
-        declaration.name.name,
-        buildInterfaceTypingContext(declaration),
-      ]),
-    ),
-    classes: new Map(
-      samlangModule.classes.map((declaration) => [
-        declaration.name.name,
-        buildClassTypingContext(moduleReference, declaration),
-      ]),
-    ),
-  };
+export function getFullyInlinedInterfaceContext(
+  instantiatedInterfaceType: SamlangIdentifierType,
+  globalTypingContext: GlobalTypingContext,
+  errorReporter: GlobalErrorReporter,
+): {
+  readonly functions: Map<string, MemberTypeInformation>;
+  readonly methods: Map<string, MemberTypeInformation>;
+  readonly superTypes: readonly SamlangIdentifierType[];
+} {
+  const interfaceTypingContext = globalTypingContext
+    .get(instantiatedInterfaceType.moduleReference)
+    ?.interfaces?.get(instantiatedInterfaceType.identifier);
+  if (interfaceTypingContext == null) {
+    return { functions: new Map(), methods: new Map(), superTypes: [] };
+  }
+  const collector: InterfaceInliningCollector = [];
+  recursiveComputeInterfaceMembersChain(
+    instantiatedInterfaceType,
+    globalTypingContext,
+    collector,
+    ModuleReferenceCollections.hashMapOf(),
+    errorReporter,
+  );
+  const functions = new Map<string, MemberTypeInformation>();
+  const methods = new Map<string, MemberTypeInformation>();
+  const superTypes: SamlangIdentifierType[] = [];
+  collector.forEach((it) => {
+    // Shadowing is allowed, as long as type matches.
+    it.functions.forEach((type, name) => functions.set(name, type));
+    it.methods.forEach((type, name) => methods.set(name, type));
+    if (it.baseInterfaceType != null) superTypes.push(it.baseInterfaceType);
+  });
+  return { functions, methods, superTypes };
+}
+
+function checkModuleMemberInterfaceConformance(
+  globalTypingContext: GlobalTypingContext,
+  actualInterface: SourceInterfaceDeclaration,
+  errorReporter: GlobalErrorReporter,
+  reportMissingMembers: boolean,
+): ReturnType<typeof getFullyInlinedInterfaceContext> {
+  const instantiatedInterfaceType = actualInterface.extendsOrImplementsNode;
+  if (instantiatedInterfaceType == null) {
+    return { functions: new Map(), methods: new Map(), superTypes: [] };
+  }
+  const fullyInlinedInterfaceContext = getFullyInlinedInterfaceContext(
+    instantiatedInterfaceType,
+    globalTypingContext,
+    errorReporter,
+  );
+
+  function checkClassMemberConformance(
+    expectedIsMethod: boolean,
+    expected: MemberTypeInformation,
+    actual: SourceClassMemberDeclaration,
+  ): void {
+    // We first filter out incompatible kind
+    if (expectedIsMethod && !actual.isMethod) {
+      errorReporter.reportUnexpectedTypeKindError(actual.location, 'method', 'function');
+      return;
+    }
+    if (!expectedIsMethod && actual.isMethod) {
+      errorReporter.reportUnexpectedTypeKindError(actual.location, 'function', 'method');
+      return;
+    }
+    const expectedTypeParameters = [...expected.typeParameters];
+    const actualTypeParameters = [...actual.typeParameters.map((it) => it.name)];
+    const expectedType = performTypeSubstitution(
+      expected.type,
+      new Map(
+        expectedTypeParameters.map((name, i) => [
+          name.name,
+          SourceIdentifierType(DummySourceReason, ModuleReference.DUMMY, `_T${i}`, []),
+        ]),
+      ),
+    );
+    const actualType = performTypeSubstitution(
+      actual.type,
+      new Map(
+        actualTypeParameters.map((name, i) => [
+          name.name,
+          SourceIdentifierType(DummySourceReason, ModuleReference.DUMMY, `_T${i}`, []),
+        ]),
+      ),
+    );
+    if (!isTheSameType(expectedType, actualType)) {
+      errorReporter.reportUnexpectedTypeError(actual.location, expectedType, actualType);
+    }
+  }
+
+  const actualMembersMap = new Map(
+    actualInterface.members.map((member) => [member.name.name, member]),
+  );
+  const missingMembers: string[] = [];
+  for (const [name, expectedMember] of fullyInlinedInterfaceContext.functions) {
+    const actualMember = actualMembersMap.get(name);
+    if (actualMember == null) {
+      missingMembers.push(name);
+      continue;
+    }
+    checkClassMemberConformance(false, expectedMember, actualMember);
+  }
+  for (const [name, expectedMember] of fullyInlinedInterfaceContext.methods) {
+    const actualMember = actualMembersMap.get(name);
+    if (actualMember == null) {
+      missingMembers.push(name);
+      continue;
+    }
+    checkClassMemberConformance(true, expectedMember, actualMember);
+  }
+  if (reportMissingMembers && missingMembers.length > 0) {
+    errorReporter.reportMissingDefinitionsError(actualInterface.location, missingMembers);
+  }
+
+  return fullyInlinedInterfaceContext;
+}
+
+function optimizeGlobalTypingContextWithInterfaceConformanceChecking(
+  sources: Sources<SamlangModule>,
+  globalTypingContext: GlobalTypingContext,
+  builtinModuleTypes: ModuleTypingContext,
+  errorReporter: GlobalErrorReporter,
+): GlobalTypingContext {
+  const optimizedGlobalTypingContext = ModuleReferenceCollections.hashMapOf<ModuleTypingContext>([
+    ModuleReference.ROOT,
+    builtinModuleTypes,
+  ]);
+  sources.forEach((samlangModule, moduleReference) => {
+    const moduleTypingContext = globalTypingContext.forceGet(moduleReference);
+    const optimizedModuleTypingContext = {
+      classes: new Map<string, ClassTypingContext>(),
+      interfaces: new Map<string, InterfaceTypingContext>(),
+    };
+    samlangModule.classes.forEach((declaration) => {
+      const fullyInlinedInterfaceContext = checkModuleMemberInterfaceConformance(
+        globalTypingContext,
+        declaration,
+        errorReporter,
+        /* reportMissingMembers */ true,
+      );
+      const optimizedClassTypingContext = {
+        ...checkNotNull(moduleTypingContext.classes.get(declaration.name.name)),
+        superTypes: fullyInlinedInterfaceContext.superTypes,
+      };
+      optimizedModuleTypingContext.classes.set(declaration.name.name, optimizedClassTypingContext);
+    });
+    samlangModule.interfaces.forEach((declaration) => {
+      const fullyInlinedInterfaceContext = checkModuleMemberInterfaceConformance(
+        globalTypingContext,
+        declaration,
+        errorReporter,
+        /* reportMissingMembers */ false,
+      );
+      const unoptimizedInterfaceTypingContext = checkNotNull(
+        moduleTypingContext.interfaces.get(declaration.name.name),
+      );
+      unoptimizedInterfaceTypingContext.functions.forEach((info, name) => {
+        fullyInlinedInterfaceContext.functions.set(name, info);
+      });
+      unoptimizedInterfaceTypingContext.methods.forEach((info, name) => {
+        fullyInlinedInterfaceContext.methods.set(name, info);
+      });
+      optimizedModuleTypingContext.interfaces.set(declaration.name.name, {
+        ...unoptimizedInterfaceTypingContext,
+        ...fullyInlinedInterfaceContext,
+      });
+    });
+    optimizedGlobalTypingContext.set(moduleReference, optimizedModuleTypingContext);
+  });
+  return optimizedGlobalTypingContext;
 }
 
 export function buildGlobalTypingContext(
   sources: Sources<SamlangModule>,
+  errorReporter: GlobalErrorReporter,
   builtinModuleTypes: ModuleTypingContext = DEFAULT_BUILTIN_TYPING_CONTEXT,
 ): GlobalTypingContext {
   const modules = ModuleReferenceCollections.hashMapOf<ModuleTypingContext>();
   modules.set(ModuleReference.ROOT, builtinModuleTypes);
+
+  function buildInterfaceTypingContext({
+    typeParameters,
+    members,
+    extendsOrImplementsNode,
+  }: SourceInterfaceDeclaration) {
+    const functions = new Map<string, MemberTypeInformation>();
+    const methods = new Map<string, MemberTypeInformation>();
+    members.forEach(({ name, isPublic, isMethod, type, typeParameters: memberTypeParameters }) => {
+      const typeInformation = {
+        isPublic,
+        typeParameters: memberTypeParameters.map((it) => ({ name: it.name.name, bound: it.bound })),
+        type,
+      };
+      if (isMethod) {
+        methods.set(name.name, typeInformation);
+      } else {
+        functions.set(name.name, typeInformation);
+      }
+    });
+    return {
+      typeParameters: typeParameters.map((it) => ({ name: it.name.name, bound: it.bound })),
+      extendsOrImplements: extendsOrImplementsNode ?? null,
+      superTypes: [],
+      functions,
+      methods,
+    };
+  }
+
+  function buildClassTypingContext(
+    moduleReference: ModuleReference,
+    classDefinition: SourceClassDefinition,
+  ): ClassTypingContext {
+    const { typeParameters, functions, methods } = buildInterfaceTypingContext(classDefinition);
+    const classType = SourceIdentifierType(
+      SourceReason(classDefinition.name.location, classDefinition.name.location),
+      moduleReference,
+      classDefinition.name.name,
+      classDefinition.typeParameters.map((it) =>
+        SourceIdentifierType(
+          SourceReason(it.location, it.location),
+          moduleReference,
+          it.name.name,
+          [],
+        ),
+      ),
+    );
+    const { typeDefinition } = classDefinition;
+    const typeDefinitionReason = SourceReason(typeDefinition.location, typeDefinition.location);
+    if (typeDefinition.type === 'object') {
+      functions.set('init', {
+        isPublic: true,
+        typeParameters,
+        type: SourceFunctionType(
+          typeDefinitionReason,
+          typeDefinition.names.map((it) => checkNotNull(typeDefinition.mappings[it.name]).type),
+          classType,
+        ),
+      });
+    } else {
+      Object.entries(typeDefinition.mappings).forEach(([tag, { type }]) => {
+        functions.set(tag, {
+          isPublic: true,
+          typeParameters,
+          type: SourceFunctionType(typeDefinitionReason, [type], classType),
+        });
+      });
+    }
+    return {
+      typeParameters,
+      typeDefinition,
+      extendsOrImplements: classDefinition.extendsOrImplementsNode ?? null,
+      superTypes: [],
+      functions,
+      methods,
+    };
+  }
+
   sources.forEach((samlangModule, moduleReference) => {
-    modules.set(moduleReference, buildModuleTypingContext(moduleReference, samlangModule));
+    modules.set(moduleReference, {
+      interfaces: new Map(
+        samlangModule.interfaces.map((declaration) => [
+          declaration.name.name,
+          buildInterfaceTypingContext(declaration),
+        ]),
+      ),
+      classes: new Map(
+        samlangModule.classes.map((declaration) => [
+          declaration.name.name,
+          buildClassTypingContext(moduleReference, declaration),
+        ]),
+      ),
+    });
   });
-  return modules;
+
+  return optimizeGlobalTypingContextWithInterfaceConformanceChecking(
+    sources,
+    modules,
+    builtinModuleTypes,
+    errorReporter,
+  );
 }
